pyproject.toml
[project]
name = "genai_mediaplan"
version = "0.1.0"
description = "genai-mediaplan using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.141.0,<1.0.0",
    "google-api-python-client>=2.176.0",
    "google-auth-oauthlib>=1.2.2",
    "mysql-connector-python>=9.3.0",
    "openai>=1.95.1",
    "PyYAML>=6.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
]

[project.scripts]
genai_mediaplan = "genai_mediaplan.main:run"
run_crew = "genai_mediaplan.main:run"
train = "genai_mediaplan.main:train"
replay = "genai_mediaplan.main:replay"
test = "genai_mediaplan.main:test"
api = "genai_mediaplan.api_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
